Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/2dcontext/drawing-rectangles-to-the-canvas/2d.fillRect.nonfinite-expected.txt

2d.fillRect.nonfinite
fillRect() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS fillRect() with Infinity/NaN is ignored 
